timeline機能の追加 ActiveDecoratorとHelperの追加
