+++ Gemfile
@@ -24,6 +24,7 @@ gem 'faker'
 # Use Active Storage variant
 gem 'image_processing', '~> 1.2'
 gem 'annotate'
+gem 'active_decorator'
 gem 'better_errors'
 gem 'binding_of_caller'
 gem 'devise'
