+++ app/assets/stylesheets/tabs.scss
@@ -12,9 +12,16 @@
     text-align: center;
     cursor: pointer;
 
+    a {
+      display: block;
+    }
+
     &.active {
       background-color: $primary-color;
-      color: white;
+
+      a {
+        color: white;
+      }
     }
 
     &:hover {
